Refactor - move common functions to deps.py
